pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "krpg"
version = "2.0.0"
description = "**KRPG** - захватывающая приключенческая игра, где вы, исследователь-путешественник, направляетесь в деревню с целью пополнения запасов. Однако прибытие в деревню сталкивает вас с загадочными событиями, заставляя вас задержаться и помочь разгадать тайны и решить проблемы. Ваш выбор остаться открывает двери к опасным приключениям, и вы сталкиваетесь с разнообразными персонажами, каждый из которых несет свои собственные секреты. Ваше путешествие станет важным моментом в судьбе деревни, выяснить происходящее"
authors = [ "Kotaz",]
license = "MIT"
readme = "README.md"

[tool.bandit]
skips = [ "B311",]

[tool.poetry.dependencies]
python = "^3.9"
rich = "^13.7.0"
msgpack = "^1.0.7"
prompt-toolkit = "^3.0.36"
urllib3 = "^2.1.0"
toml = "^0.10.2"
attrs = "^24.2.0"
questionary = "^2.0.1"

[tool.poetry.group.dev.dependencies]
types-toml = "^0.10.8.20240310"
ruff = "^0.8.1"
pyright = "^1.1.390"
